# Refactor `WorkerThread` runloop; avoid pathological starvation of pollers