--- a/core/jvm/src/main/scala/cats/effect/unsafe/WorkerThread.scala
+++ b/core/jvm/src/main/scala/cats/effect/unsafe/WorkerThread.scala
@@ -561,7 +561,7 @@ private[effect] final class WorkerThread[P <: AnyRef](
           false
         }
       } else {
-        false
+        true
       }
     }