--- a/tests/jvm/src/test/scala/cats/effect/IOPlatformSpecification.scala
+++ b/tests/jvm/src/test/scala/cats/effect/IOPlatformSpecification.scala
@@ -582,6 +582,36 @@ trait IOPlatformSpecification extends DetectPlatform { self: BaseSpec with Scala
         }
       }
 
+      "external work does not starve poll" in {
+
+        val (pool, poller, shutdown) = IORuntime.createWorkStealingComputeThreadPool(
+          threads = 1,
+          pollingSystem = DummySystem)
+
+        implicit val runtime: IORuntime =
+          IORuntime.builder().setCompute(pool, shutdown).addPoller(poller, () => ()).build()
+
+        try {
+          def mkExternalWork: Runnable = { () =>
+            val latch = new CountDownLatch(1)
+            ExecutionContext.global.execute { () =>
+              pool.execute(mkExternalWork)
+              latch.countDown()
+            }
+            latch.await() // wait until next task is in external queue
+          }
+
+          val test = IO(mkExternalWork.run()) *>
+            IO.pollers.map(_.head.asInstanceOf[DummyPoller]).flatMap { poller =>
+              poller.poll.as(true)
+            }
+
+          test.unsafeRunTimed(1.second) must beSome(beTrue)
+        } finally {
+          runtime.shutdown()
+        }
+      }
+
       if (javaMajorVersion >= 21)
         "block in-place on virtual threads" in real {
           val loomExec = classOf[Executors]