--- a/tests/jvm/src/test/scala/cats/effect/IOPlatformSpecification.scala
+++ b/tests/jvm/src/test/scala/cats/effect/IOPlatformSpecification.scala
@@ -598,7 +598,11 @@ trait IOPlatformSpecification extends DetectPlatform { self: BaseSpec with Scala
               pool.execute(mkExternalWork)
               latch.countDown()
             }
-            latch.await() // wait until next task is in external queue
+            try {
+              latch.await() // wait until next task is in external queue
+            } catch {
+              case _: InterruptedException => // ignore, runtime is shutting down
+            }
           }
 
           val test = IO(mkExternalWork.run()) *>