Fixes https://github.com/typelevel/cats-effect/issues/4228.

At a high level, a `WorkerThread` is always in one of the following three states:

1. Working (primarily on the local queue).
2. Looking for work (external or stolen).
3. Parked.

Previously, (3) was a separate `parkLoop()`, but (1) and (2) were entangled in the top-level runloop. Now, (2) is refactored into a separate `lookForWork()` loop. Thus, the top-level runloop is only responsible for tracking "ticks", such that polling runs every 64 ticks.

![worker-runloop-fsm](https://github.com/user-attachments/assets/1e432986-9777-4618-bb02-5002f724c98e)