@armanbilge 2025-01-23
Aha, all the CI failures are on JDK 21, and this is the test it's failing on.
https://github.com/typelevel/cats-effect/blob/fc5818dd512fbc7e281f91c54f9332985313d36d/tests/jvm/src/test/scala/cats/effect/IOPlatformSpecification.scala#L615-L630