@djspiewak 2025-01-25
Haven't reviewed the code yet, but I ran some benchmarks to sanity check. Looks like a very slight regression (note the `parTraverse` results in particular), but considering it fixes a number of fairness issues, I don't consider that to be an impediment.

## Before

```
[info] Benchmark                                             (cpuTokens)   (size)   Mode  Cnt     Score   Error    Units
[info] ParallelBenchmark.parTraverse                               10000     1000  thrpt   10   755.287 ± 1.087    ops/s
[info] ParallelBenchmark.traverse                                  10000     1000  thrpt   10    68.832 ± 0.112    ops/s
[info] WorkStealingBenchmark.alloc                                   N/A  1000000  thrpt   10    11.866 ± 0.089  ops/min
[info] WorkStealingBenchmark.manyThreadsSchedulingBenchmark          N/A  1000000  thrpt   10    24.042 ± 2.950  ops/min
[info] WorkStealingBenchmark.runnableScheduling                      N/A  1000000  thrpt   10  2541.428 ± 5.020  ops/min
[info] WorkStealingBenchmark.runnableSchedulingScalaGlobal           N/A  1000000  thrpt   10  1874.609 ± 6.710  ops/min
[info] WorkStealingBenchmark.scheduling                              N/A  1000000  thrpt   10    25.641 ± 3.496  ops/min
```

## After

```
[info] Benchmark                                             (cpuTokens)   (size)   Mode  Cnt     Score   Error    Units
[info] ParallelBenchmark.parTraverse                               10000     1000  thrpt   10   732.683 ± 1.429    ops/s
[info] ParallelBenchmark.traverse                                  10000     1000  thrpt   10    68.975 ± 0.100    ops/s
[info] WorkStealingBenchmark.alloc                                   N/A  1000000  thrpt   10    11.892 ± 0.135  ops/min
[info] WorkStealingBenchmark.manyThreadsSchedulingBenchmark          N/A  1000000  thrpt   10    30.569 ± 3.926  ops/min
[info] WorkStealingBenchmark.runnableScheduling                      N/A  1000000  thrpt   10  2576.645 ± 5.654  ops/min
[info] WorkStealingBenchmark.runnableSchedulingScalaGlobal           N/A  1000000  thrpt   10  1905.347 ± 2.478  ops/min
[info] WorkStealingBenchmark.scheduling                              N/A  1000000  thrpt   10    25.254 ± 4.653  ops/min
```